rename annotation package
