+++ azure-functions-java-core/src/main/java/com/microsoft/azure/serverless/functions/annotation/QueueTrigger.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.azure.serverless.functions.annotations;
+package com.microsoft.azure.serverless.functions.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -12,7 +12,11 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.PARAMETER, ElementType.METHOD})
-public @interface HttpOutput {
+@Target(ElementType.PARAMETER)
+public @interface QueueTrigger {
     String name();
+
+    String queueName();
+
+    String connection();
 }
