+++ azure-functions-java-core/src/main/java/com/microsoft/azure/serverless/functions/annotation/TimerTrigger.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.azure.serverless.functions.annotations;
+package com.microsoft.azure.serverless.functions.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -12,11 +12,13 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.PARAMETER, ElementType.METHOD})
-public @interface QueueOutput {
+@Target(ElementType.PARAMETER)
+public @interface TimerTrigger {
     String name();
 
-    String queueName();
+    String schedule();
 
-    String connection();
+    boolean useMonitor() default false;
+
+    boolean runOnStartup() default false;
 }
