--- azure-functions-java-core/src/main/java/com/microsoft/azure/serverless/functions/annotations/HttpTrigger.java
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for
- * license information.
- */
-
-package com.microsoft.azure.serverless.functions.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.PARAMETER)
-public @interface HttpTrigger {
-    enum AuthorizationLevel {
-        ANONYMOUS,
-        USER,
-        FUNCTION,
-        SYSTEM,
-        ADMIN
-    }
-
-    String name();
-
-    String route() default "";
-
-    String[] methods() default {};
-
-    AuthorizationLevel authLevel() default AuthorizationLevel.FUNCTION;
-
-    String webHookType() default "";
-}
